requirements.txt
aiogram==2.20
cog==0.0.3
dill==0.3.5.1
dlib==19.24.0
easydict==1.9
fastai==1.0.51
ffmpeg==1.4
ffmpeg_python==0.2.0
Flask==2.1.2
imageio==2.19.2
ipython==8.4.0
matplotlib==3.5.2
notebook==6.4.11
numpy==1.22.3
opencv_python==4.5.5.64
Pillow==9.1.1
PySimpleGUI==4.60.1
PyYAML==6.0
requests==2.27.1
scikit_image==0.19.2
scipy==1.8.0
sets==0.3.2
setuptools==60.10.0
# skimage==0.0
tensorboardX==2.5
tensorflow==2.9.1
torch==1.11.0
torchsummary==1.5.1
torchvision==0.12.0
tqdm==4.63.0
youtube_dl==2019.4.17
